Fix time resolution type signature

+++ dedisp/cpp/src/kernels.cpp
@@ -9,7 +9,7 @@ namespace dedisp {
 
 // TODO: update interface!?
 void fourier_domain_dedisperse(size_t dm_count, size_t n_frequencies,
-                               size_t n_channels, size_t time_resolution,
+                               size_t n_channels, float time_resolution,
                                float *spin_frequencies,
                                float *dispersion_measures, float *delays,
                                size_t stride_in, size_t stride_out,
